+++ src/collective/aviary/testing.py
@@ -1,9 +1,49 @@
 # -*- coding: utf-8 -*-
 
+from PIL import Image
+from plone.app.robotframework.testing import AUTOLOGIN_LIBRARY_FIXTURE
 from plone.app.testing import FunctionalTesting
 from plone.app.testing import IntegrationTesting
 from plone.app.testing import PLONE_FIXTURE
 from plone.app.testing import PloneSandboxLayer
+from plone.app.testing import setRoles
+from plone.app.testing import TEST_USER_ID
+from plone.testing import z2
+from StringIO import StringIO
+
+import random
+
+
+def generate_jpeg(width, height):
+    # Mandelbrot fractal
+    # FB - 201003254
+    # drawing area
+    xa = -2.0
+    xb = 1.0
+    ya = -1.5
+    yb = 1.5
+    maxIt = 25  # max iterations allowed
+    # image size
+    image = Image.new('RGB', (width, height))
+    c = complex(random.random() * 2.0 - 1.0, random.random() - 0.5)
+
+    for y in range(height):
+        zy = y * (yb - ya) / (height - 1) + ya
+        for x in range(width):
+            zx = x * (xb - xa) / (width - 1) + xa
+            z = complex(zx, zy)
+            for i in range(maxIt):
+                if abs(z) > 2.0:
+                    break
+                z = z * z + c
+            r = i % 4 * 64
+            g = i % 8 * 32
+            b = i % 16 * 16
+            image.putpixel((x, y), b * 65536 + g * 256 + r)
+
+    output = StringIO()
+    image.save(output, format='PNG')
+    return output.getvalue()
 
 
 class Fixture(PloneSandboxLayer):
@@ -19,6 +59,16 @@ class Fixture(PloneSandboxLayer):
         self.applyProfile(portal, 'collective.aviary:default')
 
 
+class RobotFixture(Fixture):
+
+    def setUpPloneSite(self, portal):
+        # Install into Plone site using portal_setup
+        super(RobotFixture, self).setUpPloneSite(portal)
+        setRoles(portal, TEST_USER_ID, ['Manager'])
+        portal.invokeFactory('Image', 'image')
+        portal.image.setImage(generate_jpeg(50, 50))
+
+
 FIXTURE = Fixture()
 
 INTEGRATION_TESTING = IntegrationTesting(
@@ -28,3 +78,8 @@ INTEGRATION_TESTING = IntegrationTesting(
 FUNCTIONAL_TESTING = FunctionalTesting(
     bases=(FIXTURE,),
     name='collective.aviary:Functional')
+
+ROBOT_FIXTURE = RobotFixture()
+ROBOT_TESTING = FunctionalTesting(
+    bases=(ROBOT_FIXTURE, AUTOLOGIN_LIBRARY_FIXTURE, z2.ZSERVER_FIXTURE),
+    name='collective.aviary:Robot')