+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 1.0a2 (unreleased)
 ------------------
 
+- Launch aviary at transform view load. [jpgimenez]
+
+- Disable portlet columns at transform view. [jpgimenez]
+
 - Simplify HTML and move editable image inside content-core. [jpgimenez]